backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ollama
huggingface-hub
langchain==0.3.26
langgraph==0.5.0
langchain-huggingface
